implement admin

+++ storefront/tags/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django.contrib.contenttypes.models import ContentType  #generic relationships
+from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.fields import GenericForeignKey
 
 
@@ -9,8 +9,6 @@ class Tag(models.Model):
 
 class TaggedItem(models.Model):
     tag = models.ForeignKey(Tag, on_delete=models.CASCADE)
-
-
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveBigIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
